aggregation-worker/requirements.txt
redis[hiredis]==5.2.1
clickhouse-connect==0.8.14
psycopg2-binary==2.9.10
pytest==8.3.4
